+++ lib/flutter_size.dart
@@ -1,3 +1,4 @@
 library flutter_size;
 
 export 'src/get_size.dart';
+export 'src/layout_body.dart';