+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.0.3
+
+* Added One more Functionality
+
+-- You can pass the widget and also can pass the width and height of the widget .
+-- This Functionality return the widget under the Layout Builder
+
+* Improve Example For Better Understanding
+
+## 0.0.2 
+
+* Added Example for packages
+
 ## 0.0.1
 
 * This is initial stable version